+++ lib/src/widgets/toolbar.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:i18n_extension/i18n_widget.dart';
 
@@ -10,14 +8,11 @@ import '../models/themes/quill_icon_theme.dart';
 import '../translations/toolbar.i18n.dart';
 import '../utils/font.dart';
 import 'controller.dart';
+import 'embeds.dart';
 import 'toolbar/arrow_indicated_button_list.dart';
-import 'toolbar/camera_button.dart';
 import 'toolbar/clear_format_button.dart';
 import 'toolbar/color_button.dart';
-import 'toolbar/formula_button.dart';
 import 'toolbar/history_button.dart';
-import 'toolbar/image_button.dart';
-import 'toolbar/image_video_utils.dart';
 import 'toolbar/indent_button.dart';
 import 'toolbar/link_style_button.dart';
 import 'toolbar/quill_font_family_button.dart';
@@ -28,13 +23,10 @@ import 'toolbar/select_alignment_button.dart';
 import 'toolbar/select_header_style_button.dart';
 import 'toolbar/toggle_check_list_button.dart';
 import 'toolbar/toggle_style_button.dart';
-import 'toolbar/video_button.dart';
 
 export 'toolbar/clear_format_button.dart';
 export 'toolbar/color_button.dart';
 export 'toolbar/history_button.dart';
-export 'toolbar/image_button.dart';
-export 'toolbar/image_video_utils.dart';
 export 'toolbar/indent_button.dart';
 export 'toolbar/link_style_button.dart';
 export 'toolbar/quill_font_size_button.dart';
@@ -43,17 +35,6 @@ export 'toolbar/select_alignment_button.dart';
 export 'toolbar/select_header_style_button.dart';
 export 'toolbar/toggle_check_list_button.dart';
 export 'toolbar/toggle_style_button.dart';
-export 'toolbar/video_button.dart';
-
-typedef OnImagePickCallback = Future<String?> Function(File file);
-typedef OnVideoPickCallback = Future<String?> Function(File file);
-typedef FilePickImpl = Future<String?> Function(BuildContext context);
-typedef WebImagePickImpl = Future<String?> Function(
-    OnImagePickCallback onImagePickCallback);
-typedef WebVideoPickImpl = Future<String?> Function(
-    OnVideoPickCallback onImagePickCallback);
-typedef MediaPickSettingSelector = Future<MediaPickSetting?> Function(
-    BuildContext context);
 
 // The default size of the icon of a button.
 const double kDefaultIconSize = 18;
@@ -69,9 +50,9 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
     this.toolbarSectionSpacing = 4,
     this.multiRowsDisplay = true,
     this.color,
-    this.filePickImpl,
     this.customButtons = const [],
     this.locale,
+    VoidCallback? afterButtonPressed,
     Key? key,
   }) : super(key: key);
 
@@ -108,19 +89,8 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
     bool showUndo = true,
     bool showRedo = true,
     bool multiRowsDisplay = true,
-    bool showImageButton = true,
-    bool showVideoButton = true,
-    bool showFormulaButton = false,
-    bool showCameraButton = true,
     bool showDirection = false,
     bool showSearchButton = true,
-    OnImagePickCallback? onImagePickCallback,
-    OnVideoPickCallback? onVideoPickCallback,
-    MediaPickSettingSelector? mediaPickSettingSelector,
-    MediaPickSettingSelector? cameraPickSettingSelector,
-    FilePickImpl? filePickImpl,
-    WebImagePickImpl? webImagePickImpl,
-    WebVideoPickImpl? webVideoPickImpl,
     List<QuillCustomButton> customButtons = const [],
 
     ///Map of font sizes in string
@@ -129,12 +99,19 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
     ///Map of font families in string
     Map<String, String>? fontFamilyValues,
 
+    /// Toolbar items to display for controls of embed blocks
+    List<EmbedButtonBuilder>? embedButtons,
+
     ///The theme to use for the icons in the toolbar, uses type [QuillIconTheme]
     QuillIconTheme? iconTheme,
 
     ///The theme to use for the theming of the [LinkDialog()],
     ///shown when embedding an image, for example
     QuillDialogTheme? dialogTheme,
+
+    /// Callback to be called after any button on the toolbar is pressed.
+    /// Is called after whatever logic the button performs has run.
+    VoidCallback? afterButtonPressed,
 
     /// The locale to use for the editor toolbar, defaults to system locale
     /// More at https://github.com/singerdmx/flutter-quill#translation
@@ -153,8 +130,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
           showColorButton ||
           showBackgroundColorButton ||
           showClearFormat ||
-          onImagePickCallback != null ||
-          onVideoPickCallback != null,
+          embedButtons?.isNotEmpty == true,
       showAlignmentButtons || showDirection,
       showLeftAlignment,
       showCenterAlignment,
@@ -197,6 +173,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
       multiRowsDisplay: multiRowsDisplay,
       customButtons: customButtons,
       locale: locale,
+      afterButtonPressed: afterButtonPressed,
       children: [
         if (showUndo)
           HistoryButton(
@@ -205,6 +182,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             undo: true,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showRedo)
           HistoryButton(
@@ -213,6 +191,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             undo: false,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showFontFamily)
           QuillFontFamilyButton(
@@ -236,6 +215,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
                   'font', newFont == 'Clear' ? null : newFont));
             },
             rawItemsMap: fontFamilies,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showFontSize)
           QuillFontSizeButton(
@@ -258,6 +238,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
                   'size', newSize == '0' ? null : getFontSize(newSize)));
             },
             rawItemsMap: fontSizes,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showBoldButton)
           ToggleStyleButton(
@@ -266,6 +247,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showItalicButton)
           ToggleStyleButton(
@@ -274,6 +256,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showSmallButton)
           ToggleStyleButton(
@@ -282,6 +265,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showUnderLineButton)
           ToggleStyleButton(
@@ -290,6 +274,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showStrikeThrough)
           ToggleStyleButton(
@@ -298,6 +283,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showInlineCode)
           ToggleStyleButton(
@@ -306,6 +292,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showColorButton)
           ColorButton(
@@ -314,6 +301,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             background: false,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showBackgroundColorButton)
           ColorButton(
@@ -322,6 +310,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             background: true,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showClearFormat)
           ClearFormatButton(
@@ -329,57 +318,11 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
-        if (showImageButton)
-          ImageButton(
-            icon: Icons.image,
-            iconSize: toolbarIconSize,
-            controller: controller,
-            onImagePickCallback: onImagePickCallback,
-            filePickImpl: filePickImpl,
-            webImagePickImpl: webImagePickImpl,
-            mediaPickSettingSelector: mediaPickSettingSelector,
-            iconTheme: iconTheme,
-            dialogTheme: dialogTheme,
-          ),
-        if (showVideoButton)
-          VideoButton(
-            icon: Icons.movie_creation,
-            iconSize: toolbarIconSize,
-            controller: controller,
-            onVideoPickCallback: onVideoPickCallback,
-            filePickImpl: filePickImpl,
-            webVideoPickImpl: webImagePickImpl,
-            mediaPickSettingSelector: mediaPickSettingSelector,
-            iconTheme: iconTheme,
-            dialogTheme: dialogTheme,
-          ),
-        if ((onImagePickCallback != null || onVideoPickCallback != null) &&
-            showCameraButton)
-          CameraButton(
-            icon: Icons.photo_camera,
-            iconSize: toolbarIconSize,
-            controller: controller,
-            onImagePickCallback: onImagePickCallback,
-            onVideoPickCallback: onVideoPickCallback,
-            filePickImpl: filePickImpl,
-            webImagePickImpl: webImagePickImpl,
-            webVideoPickImpl: webVideoPickImpl,
-            cameraPickSettingSelector: cameraPickSettingSelector,
-            iconTheme: iconTheme,
-          ),
-        if (showFormulaButton)
-          FormulaButton(
-            icon: Icons.functions,
-            iconSize: toolbarIconSize,
-            controller: controller,
-            onImagePickCallback: onImagePickCallback,
-            filePickImpl: filePickImpl,
-            webImagePickImpl: webImagePickImpl,
-            mediaPickSettingSelector: mediaPickSettingSelector,
-            iconTheme: iconTheme,
-            dialogTheme: dialogTheme,
-          ),
+        if (embedButtons != null)
+          for (final builder in embedButtons)
+            builder(controller, toolbarIconSize, iconTheme, dialogTheme),
         if (showDividers &&
             isButtonGroupShown[0] &&
             (isButtonGroupShown[1] ||
@@ -401,6 +344,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             showCenterAlignment: showCenterAlignment,
             showRightAlignment: showRightAlignment,
             showJustifyAlignment: showJustifyAlignment,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showDirection)
           ToggleStyleButton(
@@ -409,6 +353,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_textdirection_r_to_l,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showDividers &&
             isButtonGroupShown[1] &&
@@ -426,6 +371,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showDividers &&
             showHeaderStyle &&
@@ -445,6 +391,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_list_numbered,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showListBullets)
           ToggleStyleButton(
@@ -453,6 +400,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_list_bulleted,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showListCheck)
           ToggleCheckListButton(
@@ -461,6 +409,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.check_box,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showCodeBlock)
           ToggleStyleButton(
@@ -469,6 +418,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.code,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showDividers &&
             isButtonGroupShown[3] &&
@@ -485,6 +435,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_quote,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showIndent)
           IndentButton(
@@ -493,6 +444,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             isIncrease: true,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showIndent)
           IndentButton(
@@ -501,6 +453,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             isIncrease: false,
             iconTheme: iconTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showDividers && isButtonGroupShown[4] && isButtonGroupShown[5])
           VerticalDivider(
@@ -514,6 +467,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
             dialogTheme: dialogTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (showSearchButton)
           SearchButton(
@@ -522,6 +476,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             iconTheme: iconTheme,
             dialogTheme: dialogTheme,
+            afterButtonPressed: afterButtonPressed,
           ),
         if (customButtons.isNotEmpty)
           if (showDividers)
@@ -532,12 +487,14 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             ),
         for (var customButton in customButtons)
           QuillIconButton(
-              highlightElevation: 0,
-              hoverElevation: 0,
-              size: toolbarIconSize * kIconButtonFactor,
-              icon: Icon(customButton.icon, size: toolbarIconSize),
-              borderRadius: iconTheme?.borderRadius ?? 2,
-              onPressed: customButton.onTap),
+            highlightElevation: 0,
+            hoverElevation: 0,
+            size: toolbarIconSize * kIconButtonFactor,
+            icon: Icon(customButton.icon, size: toolbarIconSize),
+            borderRadius: iconTheme?.borderRadius ?? 2,
+            onPressed: customButton.onTap,
+            afterPressed: afterButtonPressed,
+          ),
       ],
     );
   }
@@ -553,8 +510,6 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
   /// Defaults to [ThemeData.canvasColor] of the current [Theme] if no color
   /// is given.
   final Color? color;
-
-  final FilePickImpl? filePickImpl;
 
   /// The locale to use for the editor toolbar, defaults to system locale
   /// More https://github.com/singerdmx/flutter-quill#translation